Create component join and create class
